feat: added separate component for cards
